# test: performance test with k6