--- a/e2e-tests/k6/.gitignore
+++ b/e2e-tests/k6/.gitignore
@@ -1,0 +1,2 @@
+node_modules
+screenshots