--- a/e2e-tests/k6/.gitignore
+++ b/e2e-tests/k6/.gitignore
@@ -1,2 +1,3 @@
 node_modules
 screenshots
+./k6