--- a/e2e-tests/k6/.gitignore
+++ b/e2e-tests/k6/.gitignore
@@ -1,3 +1,4 @@
 node_modules
 screenshots
+logs
 ./k6