--- a/e2e-tests/k6/conf/env.ts
+++ b/e2e-tests/k6/conf/env.ts
@@ -1,0 +1,15 @@
+import { EnvType } from '../types';
+
+export const env: EnvType = {
+  environments: {
+    dev: {
+      host: 'http://localhost:3000',
+    },
+    staging: {
+      host: 'https://atb-staging.api.mittatb.no',
+    },
+    prod: {
+      host: 'https://atb-prod.api.mittatb.no',
+    },
+  },
+};