--- a/e2e-tests/k6/types/index.ts
+++ b/e2e-tests/k6/types/index.ts
@@ -1,0 +1,17 @@
+export type EnvType = {
+  environments: {
+    [environment: string]: {
+      host: string;
+    };
+  };
+};
+
+export type FromLocationType = {
+  name: string;
+  quay: string;
+};
+
+export type ToLocationType = {
+  name: string;
+  quay?: string;
+};