… e2e-tests/k6/README.md
… e2e-tests/k6/yarn.lock