NOTE! I've adde a few test-ids to the travel planner in order to get precise locators for elements.

In preparation for https://github.com/AtB-AS/kundevendt/issues/19733, and to get a general feel of the performance, I've set up a browser performance test with k6 from Grafana Labs ([ref](https://grafana.com/docs/k6/latest/)).

The basic scenario now is to run a predefined number of parallell users/sessions for a given number of iterations with the simple scenario:
- Do a travel search (75% local searches and 25% with a regional location)
- Do a lookup of departures

The simple scenario will be extended further if necessary.

The metrics collected are custom metrics like time to first and last travel search suggestion are visible, and how long it takes to open the details for a travel suggestion or a departure. In addition, we get metrics based on [Core Web Vitals](https://web.dev/vitals/#core-web-vitals) from Google ([ref](https://grafana.com/docs/k6/latest/using-k6-browser/metrics/)).

Example output
```
     browser_data_received..................: 1.5 GB 18 MB/s
     browser_data_sent......................: 946 kB 11 kB/s
     browser_http_req_duration..............: avg=83.73ms  min=36µs     med=15.54ms  max=1.57s    p(95)=420.67ms p(99)=694.44ms count=1653
     browser_http_req_failed................: 0.00%  0 out of 1653
     browser_web_vital_cls..................: avg=0.020303 min=0        med=0.008681 max=0.112331 p(95)=0.085168 p(99)=0.112331 count=28  
     browser_web_vital_fcp..................: avg=1.19s    min=828.3ms  med=1.08s    max=1.78s    p(95)=1.66s    p(99)=1.75s    count=28  
     browser_web_vital_fid..................: avg=264.99µs min=99.99µs  med=250µs    max=500µs    p(95)=405µs    p(99)=481µs    count=20  
     browser_web_vital_inp..................: avg=36ms     min=24ms     med=40ms     max=48ms     p(95)=48ms     p(99)=48ms     count=20  
     browser_web_vital_lcp..................: avg=1.19s    min=828.3ms  med=1.08s    max=1.78s    p(95)=1.66s    p(99)=1.75s    count=28  
     browser_web_vital_ttfb.................: avg=177.36ms min=5.7ms    med=17.64ms  max=713.79ms p(95)=619.9ms  p(99)=692.55ms count=28  
     data_received..........................: 0 B    0 B/s
     data_sent..............................: 0 B    0 B/s
     iteration_duration.....................: avg=18.81s   min=12.45s   med=14.42s   max=39.22s   p(95)=38.75s   p(99)=39.13s   count=10  
     iterations.............................: 10     0.119875/s
     metric_assistant_details_open..........: avg=1.26s    min=1.19s    med=1.26s    max=1.29s    p(95)=1.29s    p(99)=1.29s    count=7   
     metric_assistant_firstResult...........: avg=1.34s    min=1.1s     med=1.33s    max=1.63s    p(95)=1.6s     p(99)=1.62s    count=7   
     metric_assistant_lastResult............: avg=2.21s    min=2.04s    med=2.15s    max=2.44s    p(95)=2.44s    p(99)=2.44s    count=7   
     metric_assistant_region_details_open...: avg=1.25s    min=1.25s    med=1.25s    max=1.25s    p(95)=1.25s    p(99)=1.25s    count=1   
     metric_assistant_region_firstResult....: avg=948.19ms min=948.19ms med=948.19ms max=948.19ms p(95)=948.19ms p(99)=948.19ms count=1   
     metric_assistant_region_lastResult.....: avg=2.68s    min=2.68s    med=2.68s    max=2.68s    p(95)=2.68s    p(99)=2.68s    count=1   
     metric_departures_details_open.........: avg=889.7ms  min=844.09ms med=893.24ms max=951.4ms  p(95)=932.5ms  p(99)=947.62ms count=10  
     metric_departures_show.................: avg=402.22ms min=278.9ms  med=326.6ms  max=698.19ms p(95)=651.35ms p(99)=688.83ms count=10  
     vus....................................: 1      min=1         max=3
     vus_max................................: 3      min=3         max=3
```